requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==4.0.3
bcrypt==4.3.0
boto3==1.28.67
botocore==1.31.67
certifi==2024.12.14
cffi==1.16.0
channels==4.3.1
charset-normalizer==3.4.1
click==8.2.1
confusable-homoglyphs==3.3.1
contourpy==1.3.2
coverage==7.6.12
cryptography==41.0.4
cycler==0.12.1
Django==5.1.3
django-cors-headers==4.7.0
django-redis==5.4.0
django-registration==5.2.1
django-storages==1.14.6
djangorestframework==3.15.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.58.5
h11==0.16.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.3
networkx==3.4.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.2.3
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-mock==3.14.0
python-dateutil==2.8.2
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
rectpack==0.2.2
redis==5.0.1
requests==2.32.3
s3transfer==0.7.0
scipy==1.15.3
six==1.16.0
sqlparse==0.4.4
tomli==2.2.1
trimesh==4.6.13
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
unittest-xml-reporting==3.2.0
uritemplate==4.2.0
urllib3==2.0.7
uvicorn==0.35.0
whitenoise==6.9.0
